Add a DependencyAPI to firrtl.options.Phase
This adds a DependencyAPI trait for Phase. This defines three member
functions that form the basis of defining dependencies between Phases:
prerequisites, dependents, and invalidates. Prerequisites define a set
of Phases that must run *before* some Phase. Dependents other Phases
which depend on this Phase (a way of "injecting" a prerequisite into
another Phase). Invalidates defines a function that can be called to
filter some collection of Phases to remove them from a Set (invalidate
that they ran).

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>